backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
strawberry-graphql==0.222.0
supabase==2.4.0
python-dotenv==1.0.1
pydantic==2.6.3
asyncpg==0.29.0
alembic==1.13.1
typing-extensions==4.9.0
starlette==0.36.3
click==8.1.7
SQLAlchemy==2.0.27 